requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Werkzeug
email-validator
flasgger
mysql-connector-python
Jinja2
requests
python-dotenv
pytest
pytest-flask
pytest-cov
pytest-mock
pytest-asyncio
pytest-xdist
google-generativeai